+++ lib/src/async/cdp.dart
@@ -1,0 +1,31 @@
+import 'dart:async';
+
+import 'package:appium_driver/src/common/webdriver_handler.dart';
+
+import 'package:webdriver/src/common/request_client.dart'; // ignore: implementation_imports
+
+class ChromeDevTools {
+  final AsyncRequestClient _client;
+  final AppiumWebDriverHandler _handler;
+
+  ChromeDevTools(this._client, this._handler);
+
+  /// Send CDP commands to the driver
+  /// https://chromedevtools.github.io/devtools-protocol/
+  Future<Map<String, dynamic>> execute(
+          String cmd, Map<String, dynamic> params) =>
+      _client.send(_handler.cdp.buildExecuteRequest(cmd, params),
+          _handler.cdp.parseExecuteResponse);
+
+  @override
+  int get hashCode => _client.hashCode;
+
+  @override
+  bool operator ==(other) =>
+      other is ChromeDevTools &&
+      other._client == _client &&
+      other._handler == _handler;
+
+  @override
+  String toString() => '$_handler.cdp($_client)';
+}