+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add `driver.cdp.execute` command to call ChromeDevTools command in WebView context
 - Added `-android viewmatcher` as `await driver.findElement(AppiumBy.viewmatcher('view matcher'))`
     ```dart
     import 'dart:convert';